+++ src/lsp/cobol_lsp/lsp_position.ml
@@ -90,6 +90,13 @@ let is_in_srcloc ~filename pos srcloc =
   srcloc != Srcloc.dummy &&
   is_in_lexloc pos (Srcloc.lexloc_in ~filename srcloc)
 
+let is_in_src ~filename pos = function
+  | Source_location loc ->
+      is_in_srcloc ~filename pos loc
+  | Process_parameter
+  | Process_environment ->
+      false
+
 (* --- *)
 
 (** {1 Translation of generalized source locations} *)