+++ src/lsp/cobol_preproc/preproc_diagnostics.ml
@@ -45,6 +45,7 @@ and missing_stuff =
 
 and unexpected_stuff =
   | Alphanumeric_literal
+  | Constant_literal_kind of Cobol_ptree.literal with_loc
   | Elif_compiler_directive of { suggestion: suggested_missing option }
   | Else_compiler_directive of { suggestion: suggested_missing option }
   | EndIf_compiler_directive
@@ -116,6 +117,8 @@ let pp_missing_stuff ppf = function
 let pp_unexpected_stuff ppf = function
   | Alphanumeric_literal ->
       Pretty.print ppf "alphanumeric@ literal"
+  | Constant_literal_kind _lit ->
+      Pretty.print ppf "class@ of@ literal@ for@ constant@ definition"
   | Elif_compiler_directive { suggestion } ->
       Pretty.print ppf ">>ELIF@ compiler@ directive";
       Fmt.(option (any ";@ " ++ pp_suggestion)) ppf suggestion
@@ -129,7 +132,8 @@ let pp_unexpected_stuff ppf = function
 
 let pp_unterminated ppf = function
   | If_compiler_directive _ ->
-      Pretty.print ppf ">>IF@ compiler@ directive"
+      Pretty.print ppf ">>IF@ compiler@ directive@ (no@ matching@ >>END-IF@ \
+                        found)"
   | Exec_block ->
       Pretty.print ppf "EXEC/END-EXEC@ block"
 
@@ -179,7 +183,7 @@ type warning =
       {
         loc: srcloc;
         var: Preproc_env.var with_loc;
-        prev_def_loc: Preproc_env.definition_loc;
+        prev_def_src: src;
       }
   (* | Compdir_warning of *)
   (*     { *)
@@ -260,14 +264,11 @@ let pp_warning ppf = function
       Pretty.print ppf "Undefined@ %a" pp_undefined_warning_stuff stuff
   | Unexpected_warning { stuff; _ } ->
       Pretty.print ppf "Unexpected@ %a" pp_unexpected_warning_stuff stuff
-  | Redefinition_of_env_variable { var; prev_def_loc; _ } ->
+  | Redefinition_of_env_variable { var; prev_def_src; _ } ->
       Pretty.print ppf "Redefinition@ of@ %a;@ previous@ definition@ was@ from@ \
-                        %t"
+                        %a"
         Preproc_env.VAR.pp ~&var
-        (fun ppf -> match prev_def_loc with
-           | Source_location l -> Cobol_common.Srcloc.pp_file_loc ppf l
-           | Process_parameter -> Pretty.print ppf "process@ parameters"
-           | Process_environment -> Pretty.print ppf "process@ environment")
+        Cobol_common.Srcloc.pp_src prev_def_src
 
 type diagnostics =
   {
@@ -292,6 +293,9 @@ let add_literal_diagnostics Cobol_data.Diagnostics.{ errors; _ } diags =
   List.fold_left begin fun diags error ->
     add_error (Literal_error error) diags
   end diags errors
+
+let literal_diagnostics diags =
+  add_literal_diagnostics diags none
 
 let translate ({ warnings; errors }: t) =
   let module DIAGS = Cobol_common.Diagnostics in