+++ test/cobol_parsing/test_combined_relations_parsing.ml
@@ -34,7 +34,7 @@ and fail_condition toks =
 
 let test_conditions =
   let alphanum str =
-    ALPHANUM { str; quotation = Double_quote; hexadecimal = false; runtime_repr = Native_bytes }
+    ALPHANUM (Cobol_ptree.alphanum_of_string str)
   in
   let chk descr toks cond =
     test_case descr `Quick (fun () -> check_condition toks cond)